requirements.txt
boto3
imageio
imgaug
lightning>=2.0.0
matplotlib
numpy==1.23
omegaconf
pandas
Pillow
PyYAML
scipy
setuptools
torch
torchmetrics
torchvision
tqdm
hydra-core
fastapi
tensorboard
jupyter
notebook
pydot
graphviz
pycm
git+https://github.com/garrettj403/SciencePlots